update import structure and tests for genlims submodule
