+++ bripipetools/__init__.py
@@ -6,3 +6,6 @@ from . import parsing
 
 # model depends on util and parsing
 from . import model
+
+# genlims depends on util and model
+from . import genlims
